[FLINK-16554][task][refactoring] split StreamTask
1. Extract static CheckpointingOperation and AsyncCheckpointRunnable
from StreamTask to separate files to reduce complexity.
No changes in logic or structure, just a simple refactoring.

2. Cleanup CheckpointingOperation and AsyncCheckpointRunnable:
Don't pass StreamTask as an argument

3. move registering AsyncCheckpointRunnable Closeable to its run() method
