+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/runtime/tasks/StreamTask.java
@@ -22,7 +22,6 @@ import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.api.common.accumulators.Accumulator;
 import org.apache.flink.configuration.TaskManagerOptions;
 import org.apache.flink.core.fs.CloseableRegistry;
-import org.apache.flink.core.fs.FileSystemSafetyNet;
 import org.apache.flink.metrics.Counter;
 import org.apache.flink.metrics.SimpleCounter;
 import org.apache.flink.runtime.checkpoint.CheckpointException;
@@ -30,7 +29,6 @@ import org.apache.flink.runtime.checkpoint.CheckpointFailureReason;
 import org.apache.flink.runtime.checkpoint.CheckpointMetaData;
 import org.apache.flink.runtime.checkpoint.CheckpointMetrics;
 import org.apache.flink.runtime.checkpoint.CheckpointOptions;
-import org.apache.flink.runtime.checkpoint.TaskStateSnapshot;
 import org.apache.flink.runtime.concurrent.FutureUtils;
 import org.apache.flink.runtime.execution.CancelTaskException;
 import org.apache.flink.runtime.execution.Environment;
@@ -52,7 +50,6 @@ import org.apache.flink.runtime.state.CheckpointStorageWorkerView;
 import org.apache.flink.runtime.state.CheckpointStreamFactory;
 import org.apache.flink.runtime.state.StateBackend;
 import org.apache.flink.runtime.state.StateBackendLoader;
-import org.apache.flink.runtime.state.TaskStateManager;
 import org.apache.flink.runtime.taskmanager.DispatcherThreadFactory;
 import org.apache.flink.runtime.util.ExecutorThreadFactory;
 import org.apache.flink.runtime.util.FatalExitExceptionHandler;
@@ -60,8 +57,6 @@ import org.apache.flink.streaming.api.TimeCharacteristic;
 import org.apache.flink.streaming.api.graph.StreamConfig;
 import org.apache.flink.streaming.api.graph.StreamEdge;
 import org.apache.flink.streaming.api.operators.MailboxExecutor;
-import org.apache.flink.streaming.api.operators.OperatorSnapshotFinalizer;
-import org.apache.flink.streaming.api.operators.OperatorSnapshotFutures;
 import org.apache.flink.streaming.api.operators.StreamOperator;
 import org.apache.flink.streaming.api.operators.StreamTaskStateInitializer;
 import org.apache.flink.streaming.api.operators.StreamTaskStateInitializerImpl;
@@ -77,7 +72,6 @@ import org.apache.flink.streaming.runtime.tasks.mailbox.MailboxExecutorFactory;
 import org.apache.flink.streaming.runtime.tasks.mailbox.MailboxProcessor;
 import org.apache.flink.streaming.runtime.tasks.mailbox.TaskMailbox;
 import org.apache.flink.streaming.runtime.tasks.mailbox.TaskMailboxImpl;
-import org.apache.flink.util.ExceptionUtils;
 import org.apache.flink.util.FlinkException;
 import org.apache.flink.util.Preconditions;
 import org.apache.flink.util.SerializedValue;
@@ -87,9 +81,7 @@ import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nullable;
 
-import java.io.Closeable;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.OptionalLong;
@@ -99,7 +91,6 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.atomic.AtomicReference;
 
 /**
  * Base class for all streaming tasks. A task is the unit of local processing that is deployed
@@ -928,14 +919,17 @@ public abstract class StreamTask<OUT, OP extends StreamOperator<OUT>>
 				checkpointMetaData.getCheckpointId(),
 				checkpointOptions.getTargetLocation());
 
-		CheckpointingOperation checkpointingOperation = new CheckpointingOperation(
-			this,
+		CheckpointingOperation.execute(
 			checkpointMetaData,
 			checkpointOptions,
+			checkpointMetrics,
 			storage,
-			checkpointMetrics);
-
-		checkpointingOperation.executeCheckpointing();
+			operatorChain,
+			getName(),
+			getCancelables(),
+			getAsyncOperationsThreadPool(),
+			getEnvironment(),
+			this);
 	}
 
 	// ------------------------------------------------------------------------
@@ -1053,344 +1047,11 @@ public abstract class StreamTask<OUT, OP extends StreamOperator<OUT>>
 		}
 	}
 
-	/**
-	 * This runnable executes the asynchronous parts of all involved backend snapshots for the subtask.
-	 */
-	@VisibleForTesting
-	protected static final class AsyncCheckpointRunnable implements Runnable, Closeable {
-
-		private final StreamTask<?, ?> owner;
-
-		private final Map<OperatorID, OperatorSnapshotFutures> operatorSnapshotsInProgress;
-
-		private final CheckpointMetaData checkpointMetaData;
-		private final CheckpointMetrics checkpointMetrics;
-
-		private final long asyncStartNanos;
-
-		private final AtomicReference<CheckpointingOperation.AsyncCheckpointState> asyncCheckpointState = new AtomicReference<>(
-			CheckpointingOperation.AsyncCheckpointState.RUNNING);
-
-		AsyncCheckpointRunnable(
-			StreamTask<?, ?> owner,
-			Map<OperatorID, OperatorSnapshotFutures> operatorSnapshotsInProgress,
-			CheckpointMetaData checkpointMetaData,
-			CheckpointMetrics checkpointMetrics,
-			long asyncStartNanos) {
-
-			this.owner = Preconditions.checkNotNull(owner);
-			this.operatorSnapshotsInProgress = Preconditions.checkNotNull(operatorSnapshotsInProgress);
-			this.checkpointMetaData = Preconditions.checkNotNull(checkpointMetaData);
-			this.checkpointMetrics = Preconditions.checkNotNull(checkpointMetrics);
-			this.asyncStartNanos = asyncStartNanos;
-		}
-
-		@Override
-		public void run() {
-			FileSystemSafetyNet.initializeSafetyNetForThread();
-			try {
-
-				TaskStateSnapshot jobManagerTaskOperatorSubtaskStates =
-					new TaskStateSnapshot(operatorSnapshotsInProgress.size());
-
-				TaskStateSnapshot localTaskOperatorSubtaskStates =
-					new TaskStateSnapshot(operatorSnapshotsInProgress.size());
-
-				for (Map.Entry<OperatorID, OperatorSnapshotFutures> entry : operatorSnapshotsInProgress.entrySet()) {
-
-					OperatorID operatorID = entry.getKey();
-					OperatorSnapshotFutures snapshotInProgress = entry.getValue();
-
-					// finalize the async part of all by executing all snapshot runnables
-					OperatorSnapshotFinalizer finalizedSnapshots =
-						new OperatorSnapshotFinalizer(snapshotInProgress);
-
-					jobManagerTaskOperatorSubtaskStates.putSubtaskStateByOperatorID(
-						operatorID,
-						finalizedSnapshots.getJobManagerOwnedState());
-
-					localTaskOperatorSubtaskStates.putSubtaskStateByOperatorID(
-						operatorID,
-						finalizedSnapshots.getTaskLocalState());
-				}
-
-				final long asyncEndNanos = System.nanoTime();
-				final long asyncDurationMillis = (asyncEndNanos - asyncStartNanos) / 1_000_000L;
-
-				checkpointMetrics.setAsyncDurationMillis(asyncDurationMillis);
-
-				if (asyncCheckpointState.compareAndSet(CheckpointingOperation.AsyncCheckpointState.RUNNING,
-					CheckpointingOperation.AsyncCheckpointState.COMPLETED)) {
-
-					reportCompletedSnapshotStates(
-						jobManagerTaskOperatorSubtaskStates,
-						localTaskOperatorSubtaskStates,
-						asyncDurationMillis);
-
-				} else {
-					LOG.debug("{} - asynchronous part of checkpoint {} could not be completed because it was closed before.",
-						owner.getName(),
-						checkpointMetaData.getCheckpointId());
-				}
-			} catch (Exception e) {
-				if (LOG.isDebugEnabled()) {
-					LOG.debug("{} - asynchronous part of checkpoint {} could not be completed.",
-						owner.getName(),
-						checkpointMetaData.getCheckpointId(),
-						e);
-				}
-				handleExecutionException(e);
-			} finally {
-				owner.cancelables.unregisterCloseable(this);
-				FileSystemSafetyNet.closeSafetyNetAndGuardedResourcesForThread();
-			}
-		}
-
-		private void reportCompletedSnapshotStates(
-			TaskStateSnapshot acknowledgedTaskStateSnapshot,
-			TaskStateSnapshot localTaskStateSnapshot,
-			long asyncDurationMillis) {
-
-			TaskStateManager taskStateManager = owner.getEnvironment().getTaskStateManager();
-
-			boolean hasAckState = acknowledgedTaskStateSnapshot.hasState();
-			boolean hasLocalState = localTaskStateSnapshot.hasState();
-
-			Preconditions.checkState(hasAckState || !hasLocalState,
-				"Found cached state but no corresponding primary state is reported to the job " +
-					"manager. This indicates a problem.");
-
-			// we signal stateless tasks by reporting null, so that there are no attempts to assign empty state
-			// to stateless tasks on restore. This enables simple job modifications that only concern
-			// stateless without the need to assign them uids to match their (always empty) states.
-			taskStateManager.reportTaskStateSnapshots(
-				checkpointMetaData,
-				checkpointMetrics,
-				hasAckState ? acknowledgedTaskStateSnapshot : null,
-				hasLocalState ? localTaskStateSnapshot : null);
-
-			LOG.debug("{} - finished asynchronous part of checkpoint {}. Asynchronous duration: {} ms",
-				owner.getName(), checkpointMetaData.getCheckpointId(), asyncDurationMillis);
-
-			LOG.trace("{} - reported the following states in snapshot for checkpoint {}: {}.",
-				owner.getName(), checkpointMetaData.getCheckpointId(), acknowledgedTaskStateSnapshot);
-		}
-
-		private void handleExecutionException(Exception e) {
-
-			boolean didCleanup = false;
-			CheckpointingOperation.AsyncCheckpointState currentState = asyncCheckpointState.get();
-
-			while (CheckpointingOperation.AsyncCheckpointState.DISCARDED != currentState) {
-
-				if (asyncCheckpointState.compareAndSet(
-					currentState,
-					CheckpointingOperation.AsyncCheckpointState.DISCARDED)) {
-
-					didCleanup = true;
-
-					try {
-						cleanup();
-					} catch (Exception cleanupException) {
-						e.addSuppressed(cleanupException);
-					}
-
-					Exception checkpointException = new Exception(
-						"Could not materialize checkpoint " + checkpointMetaData.getCheckpointId() + " for operator " +
-							owner.getName() + '.',
-						e);
-
-					// We only report the exception for the original cause of fail and cleanup.
-					// Otherwise this followup exception could race the original exception in failing the task.
-					try {
-						owner.getEnvironment().declineCheckpoint(checkpointMetaData.getCheckpointId(), checkpointException);
-					} catch (Exception unhandled) {
-						AsynchronousException asyncException = new AsynchronousException(unhandled);
-						owner.handleAsyncException("Failure in asynchronous checkpoint materialization", asyncException);
-					}
-
-					currentState = CheckpointingOperation.AsyncCheckpointState.DISCARDED;
-				} else {
-					currentState = asyncCheckpointState.get();
-				}
-			}
-
-			if (!didCleanup) {
-				LOG.trace("Caught followup exception from a failed checkpoint thread. This can be ignored.", e);
-			}
-		}
-
-		@Override
-		public void close() {
-			if (asyncCheckpointState.compareAndSet(
-				CheckpointingOperation.AsyncCheckpointState.RUNNING,
-				CheckpointingOperation.AsyncCheckpointState.DISCARDED)) {
-
-				try {
-					cleanup();
-				} catch (Exception cleanupException) {
-					LOG.warn("Could not properly clean up the async checkpoint runnable.", cleanupException);
-				}
-			} else {
-				logFailedCleanupAttempt();
-			}
-		}
-
-		private void cleanup() throws Exception {
-			LOG.debug(
-				"Cleanup AsyncCheckpointRunnable for checkpoint {} of {}.",
-				checkpointMetaData.getCheckpointId(),
-				owner.getName());
-
-			Exception exception = null;
-
-			// clean up ongoing operator snapshot results and non partitioned state handles
-			for (OperatorSnapshotFutures operatorSnapshotResult : operatorSnapshotsInProgress.values()) {
-				if (operatorSnapshotResult != null) {
-					try {
-						operatorSnapshotResult.cancel();
-					} catch (Exception cancelException) {
-						exception = ExceptionUtils.firstOrSuppressed(cancelException, exception);
-					}
-				}
-			}
-
-			if (null != exception) {
-				throw exception;
-			}
-		}
-
-		private void logFailedCleanupAttempt() {
-			LOG.debug("{} - asynchronous checkpointing operation for checkpoint {} has " +
-					"already been completed. Thus, the state handles are not cleaned up.",
-				owner.getName(),
-				checkpointMetaData.getCheckpointId());
-		}
-	}
-
 	public CloseableRegistry getCancelables() {
 		return cancelables;
 	}
 
 	// ------------------------------------------------------------------------
-
-	private static final class CheckpointingOperation {
-
-		private final StreamTask<?, ?> owner;
-
-		private final CheckpointMetaData checkpointMetaData;
-		private final CheckpointOptions checkpointOptions;
-		private final CheckpointMetrics checkpointMetrics;
-		private final CheckpointStreamFactory storageLocation;
-
-		private final OperatorChain<?, ?> operatorChain;
-
-		private long startSyncPartNano;
-		private long startAsyncPartNano;
-
-		// ------------------------
-
-		private final Map<OperatorID, OperatorSnapshotFutures> operatorSnapshotsInProgress;
-
-		public CheckpointingOperation(
-				StreamTask<?, ?> owner,
-				CheckpointMetaData checkpointMetaData,
-				CheckpointOptions checkpointOptions,
-				CheckpointStreamFactory checkpointStorageLocation,
-				CheckpointMetrics checkpointMetrics) {
-
-			this.owner = Preconditions.checkNotNull(owner);
-			this.checkpointMetaData = Preconditions.checkNotNull(checkpointMetaData);
-			this.checkpointOptions = Preconditions.checkNotNull(checkpointOptions);
-			this.checkpointMetrics = Preconditions.checkNotNull(checkpointMetrics);
-			this.storageLocation = Preconditions.checkNotNull(checkpointStorageLocation);
-			this.operatorChain = owner.operatorChain;
-			this.operatorSnapshotsInProgress = new HashMap<>(operatorChain.getNumberOfOperators());
-		}
-
-		public void executeCheckpointing() throws Exception {
-			startSyncPartNano = System.nanoTime();
-
-			try {
-				for (StreamOperatorWrapper<?, ?> operatorWrapper : operatorChain.getAllOperators(true)) {
-					checkpointStreamOperator(operatorWrapper.getStreamOperator());
-				}
-
-				if (LOG.isDebugEnabled()) {
-					LOG.debug("Finished synchronous checkpoints for checkpoint {} on task {}",
-						checkpointMetaData.getCheckpointId(), owner.getName());
-				}
-
-				startAsyncPartNano = System.nanoTime();
-
-				checkpointMetrics.setSyncDurationMillis((startAsyncPartNano - startSyncPartNano) / 1_000_000);
-
-				// we are transferring ownership over snapshotInProgressList for cleanup to the thread, active on submit
-				AsyncCheckpointRunnable asyncCheckpointRunnable = new AsyncCheckpointRunnable(
-					owner,
-					operatorSnapshotsInProgress,
-					checkpointMetaData,
-					checkpointMetrics,
-					startAsyncPartNano);
-
-				owner.cancelables.registerCloseable(asyncCheckpointRunnable);
-				owner.asyncOperationsThreadPool.execute(asyncCheckpointRunnable);
-
-				if (LOG.isDebugEnabled()) {
-					LOG.debug("{} - finished synchronous part of checkpoint {}. " +
-							"Alignment duration: {} ms, snapshot duration {} ms",
-						owner.getName(), checkpointMetaData.getCheckpointId(),
-						checkpointMetrics.getAlignmentDurationNanos() / 1_000_000,
-						checkpointMetrics.getSyncDurationMillis());
-				}
-			} catch (Exception ex) {
-				// Cleanup to release resources
-				for (OperatorSnapshotFutures operatorSnapshotResult : operatorSnapshotsInProgress.values()) {
-					if (null != operatorSnapshotResult) {
-						try {
-							operatorSnapshotResult.cancel();
-						} catch (Exception e) {
-							LOG.warn("Could not properly cancel an operator snapshot result.", e);
-						}
-					}
-				}
-
-				if (LOG.isDebugEnabled()) {
-					LOG.debug("{} - did NOT finish synchronous part of checkpoint {}. " +
-							"Alignment duration: {} ms, snapshot duration {} ms",
-						owner.getName(), checkpointMetaData.getCheckpointId(),
-						checkpointMetrics.getAlignmentDurationNanos() / 1_000_000,
-						checkpointMetrics.getSyncDurationMillis());
-				}
-
-				if (checkpointOptions.getCheckpointType().isSynchronous()) {
-					// in the case of a synchronous checkpoint, we always rethrow the exception,
-					// so that the task fails.
-					// this is because the intention is always to stop the job after this checkpointing
-					// operation, and without the failure, the task would go back to normal execution.
-					throw ex;
-				} else {
-					owner.getEnvironment().declineCheckpoint(checkpointMetaData.getCheckpointId(), ex);
-				}
-			}
-		}
-
-		@SuppressWarnings("deprecation")
-		private void checkpointStreamOperator(StreamOperator<?> op) throws Exception {
-				OperatorSnapshotFutures snapshotInProgress = op.snapshotState(
-						checkpointMetaData.getCheckpointId(),
-						checkpointMetaData.getTimestamp(),
-						checkpointOptions,
-						storageLocation);
-				operatorSnapshotsInProgress.put(op.getOperatorID(), snapshotInProgress);
-		}
-
-		private enum AsyncCheckpointState {
-			RUNNING,
-			DISCARDED,
-			COMPLETED
-		}
-	}
 
 	@VisibleForTesting
 	public static <OUT> RecordWriterDelegate<SerializationDelegate<StreamRecord<OUT>>> createRecordWriterDelegate(
