+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/runtime/tasks/LocalStateForwardingTest.java
@@ -105,13 +105,15 @@ public class LocalStateForwardingTest extends TestLogger {
 		OperatorID operatorID = new OperatorID();
 		snapshots.put(operatorID, osFuture);
 
-		StreamTask.AsyncCheckpointRunnable checkpointRunnable =
-			new StreamTask.AsyncCheckpointRunnable(
-				testStreamTask,
-				snapshots,
-				checkpointMetaData,
-				checkpointMetrics,
-				0L);
+		AsyncCheckpointRunnable checkpointRunnable = new AsyncCheckpointRunnable(
+			snapshots,
+			checkpointMetaData,
+			checkpointMetrics,
+			0L,
+			testStreamTask.getName(),
+			testStreamTask.getCancelables(),
+			testStreamTask.getEnvironment(),
+			testStreamTask);
 
 		checkpointRunnable.run();
